[WIP] [Docs] Bug fixes, many explanations, fix tests (#209)
* update

update

* update

* comments

Update XRETHAdminOracle.sol

* oof

* major patch

* fix test

* Update gas-report.txt

* Update OperatorDistributor.sol

* patches

* big update

* Update setup_sandbox.ts

* Update OperatorDistributor.sol

* Update gas-report.txt

* bug fix

* testing new deposit data

* works

* fix tests

* update

* move

* bug dfix

* Update OperatorDistributor.sol

* update

* Regenerated ABI (#216)

* update

* update

* demorgan's law bug

fix tests

* update

* update

* quit

* preserve and upcast, don't check and downcast

* Fix lockup amounts (#220)

* Made lockup calculations variable, added IRocketMinipoolManager

* Fixed a greater-equals

* update

* Update contracts/Operator/SuperNodeAccount.sol


---------


* Update README.md

* Update README.md

* Update SuperNodeAccount.sol

---------

+++ contracts/Oracle/XRETHAdminOracle.sol
@@ -16,26 +16,42 @@ struct MerkleProofParams {
     bytes32[][] merkleProof;
 }
 
+/**
+ * @title XRETHAdminOracle
+ * @notice An admin oracle to manage and update the total yield accrued.
+ */
 contract XRETHAdminOracle is IXRETHOracle, UpgradeableBase {
 
-    event TotalYieldAccruedUpdated(uint256 _amount);
+    event TotalYieldAccruedUpdated(int256 _amount);
 
-    uint256 internal _totalYieldAccrued;
+    int256 internal _totalYieldAccrued;
     uint256 internal _lastUpdatedTotalYieldAccrued;
 
     constructor() initializer {}
 
-    /// @dev Initializes the FundRouter contract with the specified directory address.
-    /// @param _directoryAddress The address of the directory contract.
+  /**
+     * @notice Initializes the Admin Oracle service with the specified directory address.
+     * @param _directoryAddress The address of the directory contract.
+     */
     function initializeAdminOracle(address _directoryAddress) public virtual initializer {
         super.initialize(_directoryAddress);
     }
 
-    function getTotalYieldAccrued() external view override returns (uint) {
+    /**
+     * @notice Retrieves the total yield accrued.
+     * @return The total yield accrued.
+     */
+    function getTotalYieldAccrued() external view override returns (int256) {
         return _totalYieldAccrued;
     }
 
-    function _setTotalYieldAccrued(bytes calldata _sig, uint256 _newTotalYieldAccrued, uint256 _sigTimeStamp) internal {
+    /**
+     * @notice Internal function to set the total yield accrued.
+     * @param _sig The signature.
+     * @param _newTotalYieldAccrued The new total yield accrued.
+     * @param _sigTimeStamp The timestamp of the signature.
+     */
+    function _setTotalYieldAccrued(bytes calldata _sig, int256 _newTotalYieldAccrued, uint256 _sigTimeStamp) internal {
         address recoveredAddress = ECDSA.recover(
             ECDSA.toEthSignedMessageHash(
                 keccak256(abi.encodePacked(_newTotalYieldAccrued, _sigTimeStamp, address(this), block.chainid))
@@ -54,18 +70,31 @@ contract XRETHAdminOracle is IXRETHOracle, UpgradeableBase {
         OperatorDistributor(_directory.getOperatorDistributorAddress()).resetOracleError();
     }
 
+    /**
+     * @notice Sets the total yield accrued.
+     * @param _sig The signature.
+     * @param _newTotalYieldAccrued The new total yield accrued.
+     * @param _sigTimeStamp The timestamp of the signature.
+     */
     function setTotalYieldAccrued(
         bytes calldata _sig,
-        uint256 _newTotalYieldAccrued,
+        int256 _newTotalYieldAccrued,
         uint256 _sigTimeStamp
     ) external {
         _setTotalYieldAccrued(_sig, _newTotalYieldAccrued, _sigTimeStamp);
     }
 
+    /**
+     * @notice Sets the total yield accrued and claims rewards using a Merkle proof.
+     * @param _merkleProofParams The Merkle proof parameters.
+     * @param _sig The signature.
+     * @param _newTotalYieldAccrued The new total yield accrued.
+     * @param _sigTimeStamp The timestamp of the signature.
+     */
     function setTotalYieldAccruedAndClaim(
         MerkleProofParams calldata _merkleProofParams,
         bytes calldata _sig,
-        uint256 _newTotalYieldAccrued,
+        int256 _newTotalYieldAccrued,
         uint256 _sigTimeStamp
     ) external {
         _setTotalYieldAccrued(_sig, _newTotalYieldAccrued, _sigTimeStamp);
